magisk: omnihax0r: improve su wipe survival
Hopefully this works better and negates any edge conditions that might cause bootloops.

+++ root/overlay/sbin/setup.sh
@@ -1,22 +1,59 @@
 #!/system/bin/sh
 ## INITIATE BLOOD SWEAT AND TEARS!!!
-# Allowing us to drop static scripts into rom, again
+# Magisk 23.0 su survival, never go without su
+# To ensure rom scripts always execute
 
-until [ -d /data/adb ]; do
-	sleep 0.1
+export PATH=/system/xbin:/system/bin
+
+until [ $(/system/bin/getprop init.svc.bootanim) = running ]; do
+
+    sleep 1
+
 done
 
-sleep 1
+until [ -e /data/adb ]; do
 
-if [ ! -e /data/adb/magisk/magiskinit ]; then
-	/system/bin/setprop persist.magisk.installed 0
-	until [ -e /data/adb/magisk/magiskinit ]; do
-	sleep 0.5
-	done
-	sleep 1
-	/system/bin/setprop sys.powerctl reboot
+    sleep 1
+
+done
+
+if [ ! -e /system/etc/restore/setup.tar.gz ]; then
+
+    exit 0
+
 fi
+
+sleep 2
 
 if [ -e /data/adb/magisk/magiskinit ]; then
+
 	/system/bin/setprop persist.magisk.installed 1
+
+elif [ ! -e /data/adb/magisk/magiskinit ]; then
+
+	/system/bin/blank_screen &
+
+	until [ $(/system/bin/getprop service.bootanim.exit) = 1 ] && [ $(/system/bin/getprop sys.bootstat.first_boot_completed) = 1 ]; do
+
+	    sleep 0
+
+	done
+
+	/system/bin/blank_screen &
+
+	/system/bin/setprop persist.magisk.installed 0
+
+	sleep 1
+
+	until [ -e /data/adb/magisk/magiskinit ] && [ -e /data/adb/magisk/magiskboot ] && [ -e /data/adb/magisk/magisk64 ] && [ -e /data/adb/magisk/magisk32 ]  && [ -e /data/adb/magisk/busybox ] && [ -e /data/adb/magisk/boot_patch.sh ] && [ -e /data/adb/magisk/addon.d.sh ]; do
+
+	    sleep 0
+
+	done
+
+	/system/bin/setprop persist.magisk.installed 1
+
+	/system/bin/setprop sys.powerctl reboot
+
 fi
+
